Fixing various legacy problems in Logic_Graph.

+++ logic_graph.h
@@ -26,11 +26,13 @@
 namespace SYNARMOSMA {
   class Logic_Graph: public Graph {
    private:  
+    // The main property of this class
     Propositional_System* logic;
     // This array measures the number of atomic propositions used in a given 
     // graph neighbourhood (a vertex and its neighbours)
     std::vector<int> logical_breadth;
-    int rationalize_topology();
+
+    double rationalize_topology();
     void compute_logical_breadth();
 
     virtual bool amputation(int);
